Accessors and Replacers for Design objects

Allows access to the elements which define a Design, enabling their extraction or replacement.

Usage

treatment(x, binary = FALSE, newdata = NULL, by = NULL, ...)

# S4 method for Design
treatment(x, binary = FALSE, newdata = NULL, by = NULL, ...)

treatment(x) <- value

# S4 method for Design
treatment(x) <- value

units_of_assignment(x, newdata = NULL, by = NULL)

# S4 method for Design
units_of_assignment(x, newdata = NULL, by = NULL)

units_of_assignment(x) <- value

# S4 method for Design
units_of_assignment(x) <- value

clusters(x, newdata = NULL, by = NULL)

# S4 method for Design
clusters(x, newdata = NULL, by = NULL)

clusters(x) <- value

# S4 method for Design
clusters(x) <- value

unitids(x)

# S4 method for Design
unitids(x)

unitids(x) <- value

# S4 method for Design
unitids(x) <- value

blocks(x, newdata = NULL, by = NULL)

# S4 method for Design
blocks(x, newdata = NULL, by = NULL)

blocks(x) <- value

# S4 method for Design
blocks(x) <- value

forcings(x, newdata = NULL, by = NULL)

# S4 method for Design
forcings(x, newdata = NULL, by = NULL)

forcings(x) <- value

# S4 method for Design
forcings(x) <- value

Arguments

x

Design object

binary

If FALSE (default), return a data.frame containing the named treatment variable. If TRUE and x has a formula in @dichotomy, return a data.frame containing a binary treatment variable with the name "z__". Errors on TRUE if treatment is non-binary @dichotomy is NULL. If "ifany", returns a binary treatment if possible (if treatment is already binary, or there's a valid @dichotomy), otherwise return original treatment.

newdata

Optionally an additional data.frame. If passed, and the unit of assignment variable is found in newdata, then the requested variable type for each unit of newdata is returned. See by argument if the name of the unit of assignment differs.

by

optional; named vector or list connecting names of cluster/unit of assignment variables in design to cluster/unit of assignment variables in data. Names represent variables in the Design; values represent variables in the data. Only needed if variable names differ.

...

Ignored.

value

Replacement. Either a vector/matrix of appropriate dimension, or a named data.frame if renaming variable as well.

Value

data.frame containing treatment variable

Details

For treatment(), when argument binary is FALSE, the treatment variable passed into the Design is returned as a one-column data.frame. If binary = TRUE is passed, and the Design either has a binary treatment variable, or has a dichotomy, a binary one-column data.frame will be returned. If the Design does not have access to binary treatment (non-binary treatment and no dichotomy specified), passing binary = TRUE will error.

binary = "ifany" is the most permissible; returning the dichotomized treatment variable if @dichotomy exists, otherwise returning the original treatment without error.

The one-column data.frame returned by treatment() is named as entered in the Design creation, but if a dichotomy is in the Design, the column name is "__z".