Generates a Design object with the given specifications.

Generates a randomized control treatment Design (rct_design()), or an observational Design (obs_design()), or a regression discontinuity Design (rd_design()).

Usage

rct_design(formula, data, subset = NULL, dichotomy = NULL)

rd_design(formula, data, subset = NULL, dichotomy = NULL)

obs_design(formula, data, subset = NULL, dichotomy = NULL)

Arguments

formula

defines the Design components

data

the data set.

subset

optionally subset the data before creating the Design object

dichotomy

optionally, a formula defining the dichotomy of the treatment variable if it isn't already 0/1 or logical. See details.

Value

a Design object of the requested type for use in further analysis

Details

The formula must include exactly one of unit_of_assignment(), uoa(), cluster(), or unitid() to identify the units of assignment (one or more variables). If defining an rd_design, the formula must also include a forcing() entry. The formula may optionally include a block() entry as well.

The treatment variable passed into the left-hand side of formula can either be logical, numeric, or character. If it is anything else, it is attemped to be converted to one of those (for example, factor and ordered are converted to numeric if the levels are numeric, otherwise to character. If the treatment is not logical or numeric with only values 0 and 1, in order to generate weights with ate() or ett(), the dichotomy argument must be used to identify the treatment and control groups. The Design creation functions (rct_design(), rd_design(), obs_design()) all support the dichotomy argument, or instead dichotomy can be passed to ett() and ate() directly.

The dichotomy argument should be a formula consisting of a conditional statement on both the left-hand side (identifying treatment levels associated with "treatment") and the right hand side (identifying treatment levels associated with "control"). For example, if your treatment variable was called dose, you might write:

dichotomy = dose > 250 ~ dose <= 250

The conditionals need not assign all values of treatment to control or treatment, for example, dose > 300 ~ dose < 200 does not handle 200 <= dose <= 300. Units of assignment with treatment values not assigned to either treatment or control as assumed to not recieve either, but will be maintained in the Design for proper standard error calculations.

The period (.) can be used to assign all other units of assignment. For example, we could have written the above example as

dichotomy = dose > 250 ~ .

or

dichotomy = . ~ dose <= 250

The dichotomy formula supports all Relational Operators, Logical Operators, and %in%.

Note that you can specify a conditional logic treatment in the formula (e.g. rct_design(dose > 250 ~ unitOfAssignment(...) but we would suggest instead passing the treatment variable directly and using dichotomy. Otherwise changing the dichotomy will require re-creating the Design, instead of simply using dichotomy(design) <- or passing dichotomy to ate() or ett().