Obtain Treatment from Design

When included as a predictor in a model formula where weights (via ate() or ett() being passed to the weights= argument) or a covariance adjustment model (via cov_adj() being passed to the offset= argument) are included, this will extract the treatment variable from the Design.

Usage

adopters(design = NULL, data = NULL)

Arguments

design

Optional Design. If the Design can't be identified in the model (usually because neither weights (ate() or ett()) nor a covariate adjustment model (cov_adj()) are found), the Design can be passed diretly.

data

Optional data set. By default adopters() will attempt to identify the appropriate data, if this fails (or you want to overwrite it), you can pass the data here.

Value

The treatment variable to be placed in the regression formula.

Details

While adopters() can be used in any situation, it is most useful for scenarios where the treatment variable is non-binary and the Design contains a Dichotomy. For example, say q is a 3-level ordinal treatment variable, and the binary comparison of interest is captured in dichotomy = q == 3 ~ q < 3. If you were to fit a model including q as a predictor, e.g. lm(y ~ q, ...), lm would treat q as the full ordinal variable. On the other hand, by calling lm(y ~ adopters(), weights = ate(des), ...), adopters() will generate the appropriate binary variable to allow estimation of treatment effects.