Convert Design between types

These functions convert a Design between RD, RCT and Obs types. Converting to RD requires the addition of a forcing variable (see forcing= argument), while converting from RD requires consenting to dropping the forcing variable (see loseforcing= argument).

Usage

as_rct_design(Design, ..., loseforcing = FALSE)

as_obs_design(Design, ..., loseforcing = FALSE)

as_rd_design(Design, data, ..., forcing)

Arguments

Design

Design to convert

...

No addiitonal options at present

loseforcing

Converting from RD to another Design type will error to avoid losing the forcing variable. Setting loseforcing = TRUE allows the conversion to automatically drop the forcing variable. Default FALSE.

data

Converting to an RD requires adding a forcing variable, which requires access to the original data.

forcing

Converting to an RD requires adding a forcing variable. This should be entered as the update to a formula, e.g. "~ . + forcing(forcevar)".

Value

Design of the updated type

Examples

des <- rct_design(z ~ cluster(cid1, cid2), data = simdata)
des
#> Randomized Control Trial
#> 
#>  Structure Variables 
#>  --------- --------- 
#>  Treatment z         
#>  Cluster   cid1, cid2
#> 
as_obs_design(des)
#> Observational Study
#> 
#>  Structure Variables 
#>  --------- --------- 
#>  Treatment z         
#>  Cluster   cid1, cid2
#> 
as_rd_design(des, simdata, forcing = ~ . + forcing(force))
#> Regression Discontinuity Design
#> 
#>  Structure Variables 
#>  --------- --------- 
#>  Treatment z         
#>  Cluster   cid1, cid2
#>  Forcing   force     
#> 
des2 <- rd_design(o ~ unitid(cid1, cid2) + forcing(force), data = simdata)
des2
#> Regression Discontinuity Design
#> 
#>  Structure Variables 
#>  --------- --------- 
#>  Treatment o         
#>  Unitid    cid1, cid2
#>  Forcing   force     
#> 
# as_rct_design(des2) # this will produce an error
as_rct_design(des2, loseforcing = TRUE)
#> Randomized Control Trial
#> 
#>  Structure Variables 
#>  --------- --------- 
#>  Treatment o         
#>  Unitid    cid1, cid2
#>