(Internal) Updates des@call's formula with the currently defined variable names.

Helper function to update the call with the appropriate variable names after they've been modified. Called within Design replacers.

Usage

.update_call_formula(design)

Arguments

design

A Design

Value

An updated formula

Details

It's return should be stuck into the design via des@call$formula <- .update_call_formula(des)