Linear Model for Intention To Treat

Generates a linear model object which allows proper estimation of variances accounting for the study design.

Usage

lmitt(obj, design, absorb = FALSE, ...)

# S3 method for formula
lmitt(obj, design = NULL, absorb = FALSE, ...)

# S3 method for lm
lmitt(obj, design = NULL, ...)

Arguments

obj

A formula or a lm object. See details.

design

Optional, explicitly specify the Design to be used. If the Design is specified elsewhere in the model (e.g. passed as an argument to any of ate(), ett(), cov_adj() or assigned()) it will be found automatically and does not need to be passed here as well. (If different Design objects are passed (either through the lm in weights or covariance adjustment, or through this argument), an error will be produced.) Alternatively, a formula creating a design (of the type of that would be passed as the first argument to rd_design(), rct_design(), or obs_design().

absorb

If TRUE, fixed effects are included for units of assignemnt/clusters identified in the Design. Excluded in FALSE. Default is FALSE.

...

Additional arguments passed to lm(). Ignored if obj is already an lm object.

Value

DirectAdjusted model.

Details

The first argument to lmitt, obj, specifies a linear regression model. The formula is specified exactly how it would be in lm(), with one deviation. Rather than including the treatment variable by name, the helper function assigned() should be used in its place. For example, if your outcome is y, you could call lmitt(y ~ assigned(), design = ...).

If the formula does not contain at least one assigned(), the right hand side of the formula will be treated as a series of strataification variables. This means that a formula such as y ~ x + q will be replaced with y ~ x:assigned() + q:assigned() such that lmitt will estimate the treatment effect within each level of both x and q. If assigned() is found anywhere on the right hand side of obj, this modification is not made.

Any additional arguments to lm() can be passed into .... See the help for lm() for a list of possible argument. Most commonly used in this scenario would be weights= to pass ate() or ett(), or offset= to pass cov_adj().

Alternatively, obj can be a pre-created lm object. No modification is made to the formula of the object. See the help for as.lmitt() for details of this conversion.

Note that although the Design creation functions (e.g. rct_design()) take an optional subset= argument used in the creation of the Design, this is not the same as the subset= argument passed to lm() or lmitt(). The subset= argument when creating a Design restricts the data used to generate the Design, but has no direct impact on the future lm() or lmitt() calls using that Design. (It can indirectly have an impact by excluding particular clusters/units of assignment from recieving a treatment assignment and thus complete case analysis removes them from the model.)

On the other hand, the subset= argument in lm() or lmitt() refers only to subsetting the data argument passed into lm() or lmitt().