(Internal) Compute variance blocks

(Internal) Compute variance blocks

Usage

.get_b12(x)

.get_a22_inverse(x)

.get_b22(x, ...)

.get_a11_inverse(x)

.get_b11(x, ...)

.get_a21(x)

Arguments

x

A DirectAdjusted model

...

Arguments to be passed to sandwich::meatCL

Value

.get_b12(): A \(p\times k\) matrix where \(p\) is the number of terms in the covariance model and \(k\) is the number of terms in the DirectAdjusted model

.get_a22_inverse(): A \(1\times1\) matrix

.get_b22(): A \((p+1)\times(p+1)\) matrix where the dimensions are given by the number of terms in the DirectAdjusted model (\(p\)) and an Intercept term. This should be a \(2\times2\) matrix given the dichotomous handling of treatment variables in this package and the use of the covariance model to offer the covariance adjustment.

.get_a11_inverse(): A \((p+1)\times(p+1)\) matrix where the dimensions are given by the number of terms in the covariance model (\(p\)) and an Intercept term.

.get_b11(): A \((p+1)\times(p+1)\) matrix the dimensions are given by the number of terms in the covariance model (\(p\)) and an Intercept term

.get_a12(): A \((p+1)\times2\) matrix where the number of rows are given by the number of terms in the covariance model (\(p\)) and an Intercept term, and the number of columns are given by the treatment and intercept terms in the direct adjustment model

Details

The B12 block is the covariance matrix of the cluster-level estimating equations for the covariance and direct adjustment models. It has a row for each term in the covariance model and a column for each term in the direct adjustment model. For any row that does not appear in both the experimental design and the covariance model data, its contribution to this matrix will be 0. Thus, if there is no overlap between the two datasets, this will return a matrix of 0's.

The A22 block is the diagonal element of the inverse expected Fisher Information matrix corresponding to the treatment estimate. As shown in the Details of .get_b22(), the estimating equations for a generalized linear model with a canonical link function can be written as $$\psi_i = (r_i / Var(y_i)) * (d\mu_i/d\eta_i) * x_i$$ The expected information matrix A22 is then the negative Jacobian of \(\psi_i\), which by likelihood theory is the variance-covariance matrix of \(\psi_i\): $$\psi_{i}\psi_{i}^{T} = E[(r_i / Var(y_i) * (d\mu_i/d\eta_i))^{2}] * x_ix_i' = (d\mu_i/d\eta_i)^{2} / Var(y_i) * x_ix_i'$$ Considering the whole sample, this can be expressed in matrix form as \(X'WX\). The output of this function is the inverse of the diagonal element corresponding to the treatment estimate.

The B22 block refers to a clustered variance estimate of the treatment effect estimate. The stats package offers family objects with canonical link functions, so the log-likelihood for a generalized linear model can be written in terms of the linear predictor as $$L(y_i, \beta, \phi, w_i) = w_i * (y_i * \beta'x_i - b(\beta'x_i)) / \phi + h(y_i; \phi)$$ The estimating equations \(\psi_i\) given by the score function can then be expressed as $$\psi_i = E[w_i * (y_i - \mu(\beta'x_i)) * x_i / \phi]$$ In section 4.4 of the second edition of Categorical Data Analysis, Agresti shows the derivative of the mean function with respect to the linear predictor is equivalent to the weighted variance for an observation divided by the estimate of the dispersion parameter. Thus, the above estimating equations can also be written as $$\psi_i = (r_i / Var(y_i)) * (d\mu_i/d\eta_i) * x_i$$

A matrix \(C\) of dimension \(n\times J\) is formed to indicate which unit each subject in the design belongs to, where \(J\) is the number of units at the level of treatment assignment. The treatment assignment-level estimating equations are then obtained via \(C'\psi\), where \(\psi\) is the matrix of estimating equations at the subject level.

The A11 block is the \(p\times p\) matrix corresponding to the unscaled inverse of the observed Fisher information of the covariance model. The observed information is given by the estimate of the negative Jacobian of the model's estimating equations. The unscaled version provided here divides by the number of observations used to fit the covariance model.

The B11 block is the block of the sandwich variance estimator corresponding to the variance-covariance matrix of the covariance model coefficient estimates. The estimates returned here are potentially clustered (by the clustering in the experimental design) if rows in the covariance model data also exist in the design. If there is no overlap between the two datasets, the variance-covariance matrix is estimated assuming the observations are independent.

The A21 block is the block of the sandwich variance estimator corresponding to the gradient of the direct adjustment model with respect to the covariates. Some of the information needed for this calculation is stored in the DirectAdjusted object's SandwichLayer offset. This block is the crossproduct of the prediction gradient and the gradient of the conditional mean vector for the direct adjustment model summed to the cluster level. In other words, we take this matrix to be $$\sum(d\psi_i / d\alpha) = -\sum(w_i/\phi) * (d\mu(\eta_i) / d\eta_i) * (d\upsilon(\zeta_i) / d\zeta_i) * (x_i c_i)x_i'$$ where \(\mu\) and \(\eta_i\) are the conditional mean function and linear predictor for the ith cluster in the direct adjustment model, and \(\upsilon\) and \(\zeta_i\) are the conditional mean function and linear predictor for the ith cluster in the covariance model.

References

Agresti, Alan. Categorical Data Analysis. 2003. Open WorldCat, https://nbn-resolving.org/urn:nbn:de:101:1-201502241089.